Move ExtendedRevisionCallGraph to core
